+++ CHANGELOG.md
@@ -1,4 +1,15 @@
+## 1.1.6
+
+* Added Sygic GPS Navigation app with the following actions:
+  * Open app
+  * View map action
+  * Directions with coordinates action (drive or walk)
+* Added documentation for Sygic deeplinks
+* Added unit tests covering all Sygic actions and exports
+* Updated the example app with a dedicated Sygic page and assets
+
 ## 1.1.5
+
 * Added Apple Maps app with the following actions:
   * Open app
   * View map action
@@ -10,9 +21,11 @@
 * Updated example app with an Apple Maps tab
 
 ## 1.1.4
+
 * Bump dependency deeplink_x_android version to 1.0.0 ([#Change](https://pub.dev/packages/deeplink_x_android/changelog#100))
 
 ## 1.1.3+3
+
 * Update coordinate model tests for coverage issue
 * Added Waze app with the following actions:
   * Open app
@@ -24,12 +37,15 @@
 * Updated example app with a Waze tab
 
 ## 1.1.3+2
+
 * Fixed changelog format
 
 ## 1.1.3+1
+
 * Added coordinate model tests for better reliability
 
 ## 1.1.3
+
 * Added Waze app with the following actions:
   * Open app
   * View map action
@@ -40,18 +56,20 @@
 * Updated example app with a Waze tab
 
 ## 1.1.2+1
+
 * Updated example project
   * Added splash screen with DeeplinkX logo
   * Added home screen with grid or list view of apps and stores
   * Splited main.dart to specific per app page
 * Updated readme
   * Added DeeplinkX logo
-  * Add demo link 
+  * Add demo link
 
 ## 1.1.2
+
 * Added Google Maps app with following actions:
   * Open app
-  * View coordinates action
+  * View map action
   * Search location action
   * Directions action
   * Directions with coordinates action
@@ -59,6 +77,7 @@
 * Updated example app with Google Maps tab
 
 ## 1.1.1
+
 * Added Slack app with following actions:
   * Open app
   * Open Team Action
@@ -68,6 +87,7 @@
 * Updated example app with Slack tab
 
 ## 1.1.0
+
 * Now DeeplinkX is more reliable for 100% test coverage
 * Bug Fixes
 * Added Zoom app with following actions:
@@ -77,6 +97,7 @@
 * Updated example app with Zoom tab
 
 ## 1.0.5
+
 * Added TikTok app with following actions:
   * Open app
   * Open Profile Action
@@ -86,6 +107,7 @@
 * Updated example app with TikTok tab
 
 ## 1.0.4
+
 * Added Pinterest app with following actions:
   * Open app
   * Open Profile Action
@@ -96,6 +118,7 @@
 * Updated example app with Pinterest tab
 
 ## 1.0.3
+
 * Added Twitter (X) app with following actions:
   * Open app
   * Open Profile Action
@@ -105,6 +128,7 @@
 * Updated example app with Twitter tab
 
 ## 1.0.2
+
 * Added YouTube app with following actions:
   * Open app
   * Open video action
@@ -115,6 +139,7 @@
 * Updated example app with YouTube tab
 
 ## 1.0.1
+
 * Added Facebook app with following actions:
   * Open app
   * Open profile by username action
@@ -125,27 +150,29 @@
 * Added platform support for apps for checking if app is installed
 * Fixed some bugs
 
-
 ## 1.0.0
+
 * First major release
 * More reliable deeplinks:
-  *  App Links
-  *  Universal Links
-  *  Android Intent call 
+  * App Links
+  * Universal Links
+  * Android Intent call
 * New methods:
-  *  Redirect To Store
-  *  Launch App
-  *  Is App Installed
+  * Redirect To Store
+  * Launch App
+  * Is App Installed
 * Added ability to ignore fallback to web or store when native app is not installed
 * Changed required permissions for Android:
   * Instead of adding intent filter query, we now use package query to check if app is installed
-  * Checkout each app's documentation for required permissions 
+  * Checkout each app's documentation for required permissions
 
 ## 0.1.3
+
 * Improved README
-  * Used complete GitHub links instead of relative paths 
+  * Used complete GitHub links instead of relative paths
 
 ## 0.1.2
+
 * Added support for LinkedIn deeplinks:
   * Open profile
   * Open company page
@@ -167,9 +194,11 @@
 * Added third party licenses in LICENSES
 
 ## 0.1.1
+
 * Improved readme
 
 ## 0.1.0
+
 * New Feature:
   * Added fallback to supported stores option when desired app is not installed
 * Added support for Mac App Store deeplinks:
@@ -182,6 +211,7 @@
   * Open app review page
 
 ## 0.0.3
+
 * Added support for Google Play Store deeplinks:
   * Open Play Store
   * Open app page by app package name